fix(notification): address review nits
- Extract addEntityFlags/buildEntityInput to entity_flags.go so shared
  helpers are not anchored to archive_all.go
- Add test for set>1 mutual-exclusivity guard in buildEntityInput
- Update ParseFuture godoc to note ISO dates bypass futurity checks
- Update stale ISO test date (2025-12-10) to 2099-01-01
